Add GIS file for Anaheim network
Using a force-directed graph and the supplied .pdf file
I was able to identify the links of key junctions and
highways and map these against an OSM base map.
A few junctions are missing or have different allowed
movements either due to model simplifications or changes
over the last few decades.